update et généralisation des retours API pour avoir les messages et data de manière uniforme. Ajout d'une partie administrateur dans le menu du dashboard

+++ src/features/companies/server/route.ts
@@ -7,6 +7,8 @@ import {
   UpdateCompanySchema,
 } from "../schemas";
 import prisma from "@/lib/prisma";
+import { z } from "zod";
+import { Prisma } from "@prisma/client";
 
 const app = new Hono()
   //*------------------*//
@@ -17,26 +19,77 @@ const app = new Hono()
     sessionMiddleware,
     zValidator("json", AddCompanySchema),
     async (c) => {
-      const { name, siret, country, users } = c.req.valid("json");
+      try {
+        const { name, siret, country, users } = c.req.valid("json");
 
-      const result = await prisma.company.create({
-        data: {
-          name: name,
-          siret: siret,
-          country: country,
-          users: {
-            connect: users.map((userId) => ({
-              id: userId,
-            })),
+        const result = await prisma.company.create({
+          data: {
+            name: name,
+            siret: siret,
+            country: country,
+            users: {
+              connect: users.map((userId) => ({
+                id: userId,
+              })),
+            },
           },
-        },
-      });
+        });
 
-      return c.json({
-        success: true,
-        message: `Enregistrement de l'entreprise ${result.name} réussie !`,
-        data: result,
-      });
+        return c.json({
+          success: true,
+          message: `Enregistrement de l'entreprise ${result.name} réussie !`,
+          data: result,
+        });
+      } catch (error) {
+        //Erreur provenant de la validation des données
+        if (error instanceof z.ZodError) {
+          const zodErrors = error.errors.map((e) => e.message);
+          return c.json({
+            success: false,
+            message:
+              zodErrors.length > 0
+                ? zodErrors[0]
+                : "Erreur dans la validation des données",
+            data: null,
+          });
+        }
+        // Erreur provenant de Prisma/Supabase
+        if (error instanceof Prisma.PrismaClientKnownRequestError) {
+          // Gestion des erreurs spécifiques de Prisma/Supabase
+          switch (error.code) {
+            case "P2002":
+              return c.json({
+                success: false,
+                message: "Une entrée avec ces valeurs existe déjà.",
+                data: null,
+              });
+            case "P2003":
+              return c.json({
+                success: false,
+                message: "Clé étrangère introuvable.",
+                data: null,
+              });
+            case "P2025":
+              return c.json({
+                success: false,
+                message: "Enregistrement introuvable.",
+                data: null,
+              });
+            default:
+              return c.json({
+                success: false,
+                message: `Erreur Prisma/Supabase: ${error.message}`,
+                data: null,
+              });
+          }
+        }
+        // Erreur inattendue
+        return c.json({
+          success: false,
+          message: "Une erreur inattendue s'est produite.",
+          data: null,
+        });
+      }
     }
   )
   //*------------------*//
@@ -44,25 +97,48 @@ const app = new Hono()
   //*------------------*//
   //Get all companies of the database
   .get("getAll", async (c) => {
-    const result = await prisma.company.findMany({
-      include: {
-        users: true,
-      },
-    });
-    return c.json({ data: result });
+    try {
+      const result = await prisma.company.findMany({
+        include: {
+          users: true,
+        },
+      });
+      return c.json({ success: true, message: "", data: result });
+    } catch (error) {
+      return c.json({
+        success: false,
+        message: "Error fetching companies",
+        data: null,
+      });
+    }
   })
   //Get one company by id
   .get("getById/:companyId", sessionMiddleware, async (c) => {
     const { companyId } = c.req.param();
-    const result = await prisma.company.findUnique({
-      where: {
-        id: companyId,
-      },
-      include: {
-        users: true,
-      },
-    });
-    return c.json({ data: result });
+    if (!companyId) {
+      return c.json({
+        success: false,
+        message: "Company ID is required",
+        data: null,
+      });
+    }
+    try {
+      const result = await prisma.company.findUnique({
+        where: {
+          id: companyId,
+        },
+        include: {
+          users: true,
+        },
+      });
+      return c.json({ success: true, message: "", data: result });
+    } catch (error) {
+      return c.json({
+        success: false,
+        message: "Error fetching company",
+        data: null,
+      });
+    }
   })
   //Get all companies by ids
   .post(
@@ -71,25 +147,64 @@ const app = new Hono()
     zValidator("json", GetCompaniesByIdsSchema),
     async (c) => {
       const { companyIds } = c.req.valid("json");
-      const result = await prisma.company.findMany({
-        where: {
-          id: {
-            in: companyIds,
+      if (
+        !companyIds ||
+        !Array.isArray(companyIds) ||
+        companyIds.length === 0
+      ) {
+        return c.json({
+          success: false,
+          message: "Company IDs are required",
+          data: null,
+        });
+      }
+      if (companyIds.length > 50) {
+        return c.json({
+          success: false,
+          message: "Too many company IDs provided",
+          data: null,
+        });
+      }
+      if (companyIds.some((id) => typeof id !== "string")) {
+        return c.json({
+          success: false,
+          message: "Invalid Company IDs",
+          data: null,
+        });
+      }
+
+      try {
+        const result = await prisma.company.findMany({
+          where: {
+            id: {
+              in: companyIds,
+            },
           },
-        },
-        include: {
-          users: true,
-        },
-      });
-      return c.json({ data: result });
+          include: {
+            users: true,
+          },
+        });
+        return c.json({ success: true, message: "", data: result });
+      } catch (error) {
+        return c.json({
+          success: false,
+          message: "Error fetching companies",
+          data: null,
+        });
+      }
     }
   )
   //Get List of companies by user id
   .get("getByUserId/:userId", async (c) => {
     const { userId } = c.req.param();
     if (!userId) {
-      return c.json({ error: "Missing userId" }, 400);
+      return c.json({
+        success: false,
+        message: "User ID is required",
+        data: null,
+      });
     }
+
     try {
       const result = await prisma.company.findMany({
         where: {
@@ -103,11 +218,13 @@ const app = new Hono()
           users: true,
         },
       });
-      return c.json({ data: result });
-    } catch (error: any) {
-      console.error("Error stack:", error?.stack);
-      console.error("Error message:", error?.message || error);
-      return c.json({ error: error?.message || "Unknown error" }, 500);
+      return c.json({ success: true, message: "", data: result });
+    } catch (error) {
+      return c.json({
+        success: false,
+        message: "Error fetching companies",
+        data: null,
+      });
     }
   })
   //*------------------*//
@@ -119,27 +236,87 @@ const app = new Hono()
     zValidator("json", UpdateCompanySchema),
     async (c) => {
       const { companyId } = c.req.param();
-      const { name, siret, country, users } = c.req.valid("json");
+      if (!companyId) {
+        return c.json({
+          success: false,
+          message: "Company ID is required",
+          data: null,
+        });
+      }
 
-      const result = await prisma.company.update({
-        where: {
-          id: companyId,
-        },
-        data: {
-          name: name,
-          siret: siret,
-          country: country,
-          users: {
-            set: users.map((userId) => ({
-              id: userId,
-            })),
+      try {
+        const { name, siret, country, users } = c.req.valid("json");
+
+        const result = await prisma.company.update({
+          where: {
+            id: companyId,
           },
-        },
-      });
-      return c.json({
-        success: true,
-        message: `Entreprise ${result.name} modifiée avec succès !`,
-      });
+          data: {
+            name: name,
+            siret: siret,
+            country: country,
+            users: {
+              set: users.map((userId) => ({
+                id: userId,
+              })),
+            },
+          },
+        });
+        return c.json({
+          success: true,
+          message: `Entreprise ${result.name} modifiée avec succès !`,
+          data: result,
+        });
+      } catch (error) {
+        //Erreur provenant de la validation des données
+        if (error instanceof z.ZodError) {
+          const zodErrors = error.errors.map((e) => e.message);
+          return c.json({
+            success: false,
+            message:
+              zodErrors.length > 0
+                ? zodErrors[0]
+                : "Erreur dans la validation des données",
+            data: null,
+          });
+        }
+        // Erreur provenant de Prisma/Supabase
+        if (error instanceof Prisma.PrismaClientKnownRequestError) {
+          // Gestion des erreurs spécifiques de Prisma/Supabase
+          switch (error.code) {
+            case "P2002":
+              return c.json({
+                success: false,
+                message: "Une entrée avec ces valeurs existe déjà.",
+                data: null,
+              });
+            case "P2003":
+              return c.json({
+                success: false,
+                message: "Clé étrangère introuvable.",
+                data: null,
+              });
+            case "P2025":
+              return c.json({
+                success: false,
+                message: "Enregistrement introuvable.",
+                data: null,
+              });
+            default:
+              return c.json({
+                success: false,
+                message: `Erreur Prisma/Supabase: ${error.message}`,
+                data: null,
+              });
+          }
+        }
+        // Erreur inattendue
+        return c.json({
+          success: false,
+          message: "Une erreur inattendue s'est produite.",
+          data: null,
+        });
+      }
     }
   )
   //*------------------*//
@@ -147,15 +324,32 @@ const app = new Hono()
   //*------------------*//
   .delete("delete/:companyId", sessionMiddleware, async (c) => {
     const { companyId } = c.req.param();
-    const result = await prisma.company.delete({
-      where: {
-        id: companyId,
-      },
-    });
-    return c.json({
-      success: true,
-      message: `Entreprise ${result.name} supprimée avec succès !`,
-    });
+    if (!companyId) {
+      return c.json({
+        success: false,
+        message: "Company ID is required",
+        data: null,
+      });
+    }
+
+    try {
+      const result = await prisma.company.delete({
+        where: {
+          id: companyId,
+        },
+      });
+      return c.json({
+        success: true,
+        message: `Entreprise ${result.name} supprimée avec succès !`,
+        data: result,
+      });
+    } catch (error) {
+      return c.json({
+        success: false,
+        message: "Erreur lors de la suppression de l'entreprise. Suppression interrompue.",
+        data: null,
+      });
+    }
   });
 
 export default app;
